updating the function for edge cases and created new routes in express to handle stats information

+++ central-service/models/statistics.go
@@ -1,9 +1,9 @@
 package models
 
 type Statistics struct {
-	UserID     string `gorm:"primaryKey" json:"id"`
-	User       User   `gorm:"foreignKey:UserID;references:ID"`
-	RoomID     string `gorm:"primaryKey" json:"shortCode"`
-	Room       Room   `gorm:"foreignKey:RoomID;references:ShortCode"`
-	TotalScore uint   `json:"score"`
+	UserID     string  `gorm:"primaryKey" json:"id"`
+	User       User    `gorm:"foreignKey:UserID;references:ID"`
+	RoomID     string  `gorm:"primaryKey" json:"shortCode"`
+	Room       Room    `gorm:"foreignKey:RoomID;references:ShortCode"`
+	TotalScore float64 `json:"score"`
 }
